+++ apps/www-mosh-lol/src/styles/index.css
@@ -1,0 +1,1 @@
+@import url("https://fonts.googleapis.com/css2?family=Press+Start+2P&display=swap");
