--- components/atoms/src/index.scss
@@ -1,4 +1,0 @@
-@font-face {
-  font-family: "AtariClassic";
-  src: local("AtariClassic"), url("../../utils/fonts/AtariClassic/AtariClassic.ttf") format("truetype");
-}
