+++ components/atoms/src/index.ts
@@ -1,5 +1,3 @@
-import "./index.scss";
-
 export { Button } from "./Button";
 export type { ButtonProps } from "./Button";
 export { ButtonLink } from "./ButtonLink";
