--- components/utils/object.ts
@@ -1,15 +1,0 @@
-export type Dict<T = any> = Record<string, T>;
-
-export const omit = <T extends Dict, K extends keyof T>(
-  object: T,
-  keys: K[],
-) => {
-  const result: Dict = {};
-
-  Object.keys(object).forEach(key => {
-    if (keys.includes(key as K)) return;
-    result[key] = object[key];
-  });
-
-  return result as Omit<T, K>;
-};
