+++ configurations/tailwindcss/src/index.js
@@ -74,6 +74,9 @@ module.exports = {
         violet: colors.violet,
         warmGray: colors.warmGray,
       },
+      fontFamily: {
+        anime: ["Press Start 2P", "cursive"],
+      },
       keyframes: {
         "fade-in-down": {
           "0%": {
@@ -121,9 +124,6 @@ module.exports = {
           colors: theme("colors"),
         };
       },
-    },
-    fontFamily: {
-      Atari: ["AtariClassic"],
     },
   },
   variants: {
